refactor(web): split api/types.ts into a per-domain types/ package
lib/api/types.ts was a single ~1,010-line file holding 92 exported
interfaces/types across every API domain. Split them into a per-domain
types/ package (repository, pages, graph, git, decisions, workspace,
ownership, intelligence, chat, etc.). types.ts now re-exports every
module, so all 46 '@/lib/api/types' importers keep resolving unchanged.

Introduce a GraphLike<N> generic ({ nodes: N[]; links: GraphEdgeResponse[] })
and have the four graph-response shapes (GraphExportResponse,
EgoGraphResponse, DeadCodeGraphResponse, HotFilesGraphResponse) build on
it instead of repeating nodes/links. Structurally identical to before.

Verified: tsc --noEmit clean, next build succeeds, eslint clean; all 92
original exports preserved (+ GraphLike).
